feat(get web service metrics):CloudWatch metrics job
The AWS SDK version has been updated in the go.mod file. In addition, a new job has been added to list CloudWatch metrics for an AWS ECS web service. The new job retrieves CPU and memory utilization metrics from CloudWatch for a given ECS service and sends the details to the chat output if the `ShouldSendChatOutput` flag is set in the job parameters.